Moved logic into controller
